+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Add support for the Framework Laptop 16.
 - Change the Intel graphics driver to allow Xe to work.
 - Change Heroic Games Launcher to be installed as a Flatpak.
+- Change Lutris to be installed as a Flatpak.
 
 ## 4.0.0 - The Conversion Update